Added in custom error codes besides 500

+++ backend/app/endpoints/room_api.py
@@ -1,14 +1,16 @@
-from flask import jsonify, request
-from flask_restful import Resource, reqparse, abort
+from flask import jsonify, request, make_response
+from flask_restful import Resource, reqparse
 from flask_restful_swagger import swagger
+from werkzeug.exceptions import BadRequest
 from typing import List
 from app import db
 from app.database.room import Room
+from app.endpoints.utils import create_400_error, create_404_error, create_500_error
 
 class RoomListApi(Resource):
     def __init__(self):
         self.reqparse = reqparse.RequestParser()
-        self.reqparse.add_argument("name", type=str, required=False, default="", location="json")
+        self.reqparse.add_argument("name", type=str, required=True, location="json")
         self.reqparse.add_argument("video_id", type=str, required=False, default="", location="json")
         super(RoomListApi, self).__init__()
 
@@ -30,7 +32,7 @@ class RoomListApi(Resource):
         try:
             return jsonify(self.__get_all_rooms())
         except:
-            abort(500)
+            return create_500_error()
     def __get_all_rooms(self):
         return [room.to_json() for room in Room.query.all()]
 
@@ -59,6 +61,9 @@ class RoomListApi(Resource):
                 "code": 200,
                 "message": "Created the new room successfully"
             }, {
+                "code": 400,
+                "message": "Bad request",
+            }, {
                 "code": 500,
                 "message": "Internal server error",
             }
@@ -68,8 +73,10 @@ class RoomListApi(Resource):
         try:
             args = self.reqparse.parse_args()
             return jsonify(self.__create_room(args["name"], args["video_id"]))
+        except BadRequest:
+            return create_400_error()
         except:
-            abort(500)
+            return create_500_error()
     def __create_room(self, room_name, video_id):
         room = Room(name=room_name, video_id=video_id)
         db.session.add(room)
@@ -79,8 +86,8 @@ class RoomListApi(Resource):
 class RoomApi(Resource):
     def __init__(self):
         self.reqparse = reqparse.RequestParser()
-        self.reqparse.add_argument("name", type=str, required=False, default="", location="json")
-        self.reqparse.add_argument("video_id", type=str, required=False, default="", location="json")
+        self.reqparse.add_argument("name", type=str, required=True, location="json")
+        self.reqparse.add_argument("video_id", type=str, required=True, location="json")
         super(RoomApi, self).__init__()
 
     @swagger.operation(
@@ -101,6 +108,9 @@ class RoomApi(Resource):
                 "code": 200,
                 "message": "Returned the room"
             }, {
+                "code": 404,
+                "message": "Resource not found",
+            }, {
                 "code": 500,
                 "message": "Internal server error",
             }
@@ -109,10 +119,15 @@ class RoomApi(Resource):
     def get(self, room_id):
         try:
             return jsonify(self.__get_room(room_id))
+        except LookupError:
+            return create_404_error()
         except:
-            abort(500)
+            return create_500_error()
     def __get_room(self, room_id):
-        return Room.query.get(room_id).to_json()
+        room = Room.query.get(room_id)
+        if room is None:
+            raise LookupError("Room not found")
+        return room.to_json()
 
     @swagger.operation(
         notes="Updates the specific room",
@@ -129,7 +144,7 @@ class RoomApi(Resource):
             {
                 "name": "name",
                 "description": "Name to update room with",
-                "required": False,
+                "required": True,
                 "allowMultiple": False,
                 "dataType": "string",
                 "paramType": "body"
@@ -137,7 +152,7 @@ class RoomApi(Resource):
             {
                 "name": "video_id",
                 "description": "ID of video to update room with",
-                "required": False,
+                "required": True,
                 "allowMultiple": False,
                 "dataType": "string",
                 "paramType": "body"
@@ -148,6 +163,12 @@ class RoomApi(Resource):
                 "code": 200,
                 "message": "Updated the room"
             }, {
+                "code": 400,
+                "message": "Bad request",
+            }, {
+                "code": 404,
+                "message": "Resource not found",
+            }, {
                 "code": 500,
                 "message": "Internal server error",
             }
@@ -156,13 +177,18 @@ class RoomApi(Resource):
     def put(self, room_id):
         try:
             return self.__update_room(room_id, self.reqparse.parse_args())
+        except (AttributeError, BadRequest):
+            return create_400_error()
+        except LookupError:
+            return create_404_error()
         except:
-            abort(500)
+            return create_500_error()
     def __update_room(self, room_id, args):
         room = Room.query.get(room_id)
+        if room is None:
+            raise LookupError("Room not found")
         for k, v in args.items():
-            if v is not None:
-                setattr(room, k, v)
+            setattr(room, k, v)
         db.session.commit()
         return room.to_json()
 
@@ -183,6 +209,9 @@ class RoomApi(Resource):
                 "code": 200,
                 "message": "Deleted the room"
             }, {
+                "code": 404,
+                "message": "Resource not found",
+            }, {
                 "code": 500,
                 "message": "Internal server error",
             }
@@ -192,10 +221,14 @@ class RoomApi(Resource):
         try:
             self.__delete_room(room_id)
             return jsonify(success=True)
+        except LookupError:
+            return create_404_error()
         except:
-            abort(500)
+            return create_500_error()
     def __delete_room(self, room_id):
         room = Room.query.get(room_id)
+        if room is None:
+            raise LookupError("Room not found")
         db.session.delete(room)
         db.session.commit()
         return
